Restores compatiblity with older ShutdownSignalBarrier.
The newer ShutdownSignalBarrier is AutoClosable and can be used in
a try-with-resources. The problem is that the older
ShutdownSignalBarrier isn't.

So when building benchmarks with an older version of Aeron/Agrona,
then this class doesn't compile and I can't make a build.

It is important that the same benchmarks can be used with different
Aeron/Agrona versions, so that when benchmarking performance differences
only Aeron/Agrona version differences matter, the benchmark version
should remain constant.